renaming package to multitenancy
