+++ spring-boot-jpa-multitenancy/schema/src/main/java/com/example/multitenancy/schema/config/WebMvcConfig.java
@@ -1,6 +1,6 @@
 package com.example.multitenancy.schema.config;
 
-import com.example.multitenancy.schema.config.tenant.TenantInterceptor;
+import com.example.multitenancy.schema.config.multitenancy.TenantInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
