+++ spring-boot-jpa-multitenancy/schema/src/main/java/com/example/multitenancy/schema/config/multitenancy/TenantInterceptor.java
@@ -1,4 +1,4 @@
-package com.example.multitenancy.schema.config.tenant;
+package com.example.multitenancy.schema.config.multitenancy;
 
 import static org.springframework.http.HttpStatus.FORBIDDEN;
