+++ spring-boot-jpa-multitenancy/schema/src/main/java/com/example/multitenancy/schema/config/multitenancy/package-info.java
@@ -1,0 +1,2 @@
+@org.springframework.lang.NonNullApi
+package com.example.multitenancy.schema.config.multitenancy;
