--- spring-boot-jpa-multitenancy/schema/src/main/java/com/example/multitenancy/schema/config/tenant/package-info.java
@@ -1,2 +1,0 @@
-@org.springframework.lang.NonNullApi
-package com.example.multitenancy.schema.config.tenant;
