+++ spring-boot-jpa-multitenancy/schema/src/main/java/com/example/multitenancy/schema/mapper/CustomerMapper.java
@@ -4,17 +4,22 @@ import com.example.multitenancy.schema.domain.request.CustomerDto;
 import com.example.multitenancy.schema.entities.Customer;
 import java.util.List;
 import org.mapstruct.InheritInverseConfiguration;
+import org.mapstruct.IterableMapping;
 import org.mapstruct.Mapper;
+import org.mapstruct.NullValueCheckStrategy;
+import org.mapstruct.NullValueMappingStrategy;
 
-@Mapper(componentModel = "spring")
+@Mapper(componentModel = "spring", nullValueCheckStrategy = NullValueCheckStrategy.ALWAYS)
 public interface CustomerMapper {
 
-    CustomerDto entityToDTO(Customer project);
+    CustomerDto entityToDTO(Customer customer);
 
+    @IterableMapping(nullValueMappingStrategy = NullValueMappingStrategy.RETURN_DEFAULT)
     List<CustomerDto> entityToDTO(Iterable<Customer> customerEntityList);
 
     @InheritInverseConfiguration
-    Customer dtoToEntity(CustomerDto project);
+    Customer dtoToEntity(CustomerDto customerDto);
 
+    @IterableMapping(nullValueMappingStrategy = NullValueMappingStrategy.RETURN_DEFAULT)
     List<Customer> dtoToEntity(Iterable<CustomerDto> customerDtoList);
 }
